Extract field analysis module from lib.rs
