+++ crudcrate-derive/src/lib.rs
@@ -1,6 +1,7 @@
 mod attribute_parser;
 mod attributes;
 mod codegen;
+mod field_analysis;
 mod field_analyzer;
 mod macro_implementation;
 mod relation_validator;
@@ -31,20 +32,6 @@ fn extract_active_model_type(input: &DeriveInput, name: &syn::Ident) -> proc_mac
     } else {
         let ident = format_ident!("{}ActiveModel", name);
         quote! { #ident }
-    }
-}
-
-fn extract_named_fields(
-    input: &DeriveInput,
-) -> syn::punctuated::Punctuated<syn::Field, syn::token::Comma> {
-    if let Data::Struct(data) = &input.data {
-        if let Fields::Named(named) = &data.fields {
-            named.named.clone()
-        } else {
-            panic!("ToCreateModel only supports structs with named fields");
-        }
-    } else {
-        panic!("ToCreateModel can only be derived for structs");
     }
 }
 
@@ -119,177 +106,6 @@ fn generate_excluded_merge_code(
             }
         })
         .collect()
-}
-
-fn extract_entity_fields(
-    input: &DeriveInput,
-) -> Result<&syn::punctuated::Punctuated<syn::Field, syn::token::Comma>, TokenStream> {
-    match &input.data {
-        Data::Struct(data) => match &data.fields {
-            Fields::Named(fields) => Ok(&fields.named),
-            _ => Err(syn::Error::new_spanned(
-                input,
-                "EntityToModels only supports structs with named fields",
-            )
-            .to_compile_error()
-            .into()),
-        },
-        _ => Err(
-            syn::Error::new_spanned(input, "EntityToModels only supports structs")
-                .to_compile_error()
-                .into(),
-        ),
-    }
-}
-
-fn parse_entity_attributes(input: &DeriveInput, struct_name: &syn::Ident) -> (syn::Ident, String) {
-    let mut api_struct_name = None;
-    let mut active_model_path = None;
-
-    for attr in &input.attrs {
-        if attr.path().is_ident("crudcrate")
-            && let Meta::List(meta_list) = &attr.meta
-            && let Ok(metas) =
-                Punctuated::<Meta, Comma>::parse_terminated.parse2(meta_list.tokens.clone())
-        {
-            for meta in &metas {
-                if let Meta::NameValue(nv) = meta {
-                    if nv.path.is_ident("api_struct") {
-                        if let syn::Expr::Lit(expr_lit) = &nv.value
-                            && let Lit::Str(s) = &expr_lit.lit
-                        {
-                            api_struct_name = Some(format_ident!("{}", s.value()));
-                        }
-                    } else if nv.path.is_ident("active_model")
-                        && let syn::Expr::Lit(expr_lit) = &nv.value
-                        && let Lit::Str(s) = &expr_lit.lit
-                    {
-                        active_model_path = Some(s.value());
-                    }
-                }
-            }
-        }
-    }
-
-    let table_name = attribute_parser::extract_table_name(&input.attrs)
-        .unwrap_or_else(|| struct_name.to_string());
-    let api_struct_name =
-        api_struct_name.unwrap_or_else(|| format_ident!("{}", table_name.to_pascal_case()));
-    let active_model_path = active_model_path.unwrap_or_else(|| "ActiveModel".to_string());
-
-    (api_struct_name, active_model_path)
-}
-
-fn analyze_entity_fields(
-    fields: &syn::punctuated::Punctuated<syn::Field, syn::token::Comma>,
-) -> EntityFieldAnalysis<'_> {
-    let mut analysis = EntityFieldAnalysis {
-        db_fields: Vec::new(),
-        non_db_fields: Vec::new(),
-        primary_key_field: None,
-        sortable_fields: Vec::new(),
-        filterable_fields: Vec::new(),
-        fulltext_fields: Vec::new(),
-        join_on_one_fields: Vec::new(),
-        join_on_all_fields: Vec::new(),
-        // join_configs: std::collections::HashMap::new(), // Removed due to HashMap key issues
-    };
-
-    for field in fields {
-        let is_non_db = attribute_parser::get_crudcrate_bool(field, "non_db_attr").unwrap_or(false);
-
-        // Check for join attributes regardless of db/non_db status
-        if let Some(join_config) = get_join_config(field) {
-            if join_config.on_one {
-                analysis.join_on_one_fields.push(field);
-            }
-            if join_config.on_all {
-                analysis.join_on_all_fields.push(field);
-            }
-            // Note: join_configs removed to avoid HashMap key issues with syn::Field
-        }
-
-        if is_non_db {
-            analysis.non_db_fields.push(field);
-        } else {
-            analysis.db_fields.push(field);
-
-            if attribute_parser::field_has_crudcrate_flag(field, "primary_key") {
-                analysis.primary_key_field = Some(field);
-            }
-            if attribute_parser::field_has_crudcrate_flag(field, "sortable") {
-                analysis.sortable_fields.push(field);
-            }
-            if attribute_parser::field_has_crudcrate_flag(field, "filterable") {
-                analysis.filterable_fields.push(field);
-            }
-            if attribute_parser::field_has_crudcrate_flag(field, "fulltext") {
-                analysis.fulltext_fields.push(field);
-            }
-        }
-    }
-
-    analysis
-}
-
-fn validate_field_analysis(analysis: &EntityFieldAnalysis) -> Result<(), TokenStream> {
-    if analysis.primary_key_field.is_some()
-        && analysis
-            .db_fields
-            .iter()
-            .filter(|field| attribute_parser::field_has_crudcrate_flag(field, "primary_key"))
-            .count()
-            > 1
-    {
-        return Err(syn::Error::new_spanned(
-            analysis.primary_key_field.unwrap(),
-            "Only one field can be marked with 'primary_key' attribute",
-        )
-        .to_compile_error()
-        .into());
-    }
-
-    // Validate that non_db_attr fields have #[sea_orm(ignore)]
-    for field in &analysis.non_db_fields {
-        if !has_sea_orm_ignore(field) {
-            let field_name = field
-                .ident
-                .as_ref()
-                .map_or_else(|| "unknown".to_string(), std::string::ToString::to_string);
-            return Err(syn::Error::new_spanned(
-                field,
-                format!(
-                    "Field '{field_name}' has #[crudcrate(non_db_attr)] but is missing #[sea_orm(ignore)].\n\
-                     Non-database fields must be marked with both attributes.\n\
-                     Add #[sea_orm(ignore)] above the #[crudcrate(...)] attribute."
-                ),
-            )
-            .to_compile_error()
-            .into());
-        }
-    }
-
-    Ok(())
-}
-
-/// Check if a field has the `#[sea_orm(ignore)]` attribute
-fn has_sea_orm_ignore(field: &syn::Field) -> bool {
-    for attr in &field.attrs {
-        if attr.path().is_ident("sea_orm")
-            && let Meta::List(meta_list) = &attr.meta
-            && let Ok(metas) =
-                Punctuated::<Meta, Comma>::parse_terminated.parse2(meta_list.tokens.clone())
-        {
-            for meta in metas {
-                if let Meta::Path(path) = meta
-                    && path.is_ident("ignore")
-                {
-                    return true;
-                }
-            }
-        }
-    }
-    false
 }
 
 fn resolve_join_field_type_preserving_container(
@@ -593,7 +409,7 @@ pub fn to_create_model(input: TokenStream) -> TokenStream {
     let create_name = format_ident!("{}Create", name);
 
     let active_model_type = extract_active_model_type(&input, name);
-    let fields = extract_named_fields(&input);
+    let fields = field_analysis::extract_named_fields(&input);
     let create_struct_fields = codegen::models::create::generate_create_struct_fields(&fields);
     let conv_lines = codegen::models::create::generate_create_conversion_lines(&fields);
 
@@ -652,7 +468,7 @@ pub fn to_update_model(input: TokenStream) -> TokenStream {
     let update_name = format_ident!("{}Update", name);
 
     let active_model_type = extract_active_model_type(&input, name);
-    let fields = extract_named_fields(&input);
+    let fields = field_analysis::extract_named_fields(&input);
     let included_fields = crate::codegen::models::update::filter_update_fields(&fields);
     let update_struct_fields =
         crate::codegen::models::update::generate_update_struct_fields(&included_fields);
@@ -730,7 +546,7 @@ pub fn to_list_model(input: TokenStream) -> TokenStream {
     let name = &input.ident;
     let list_name = format_ident!("{}List", name);
 
-    let fields = extract_named_fields(&input);
+    let fields = field_analysis::extract_named_fields(&input);
     let list_struct_fields = crate::codegen::models::list::generate_list_struct_fields(&fields);
     let list_from_assignments =
         crate::codegen::models::list::generate_list_from_assignments(&fields);
@@ -923,7 +739,7 @@ fn parse_and_validate_entity_attributes(
     input: &DeriveInput,
     struct_name: &syn::Ident,
 ) -> Result<(String, syn::Ident, String, CRUDResourceMeta), TokenStream> {
-    let (api_struct_name, active_model_path) = parse_entity_attributes(input, struct_name);
+    let (api_struct_name, active_model_path) = field_analysis::parse_entity_attributes(input, struct_name);
     let table_name = attribute_parser::extract_table_name(&input.attrs)
         .unwrap_or_else(|| struct_name.to_string());
 
@@ -1008,7 +824,7 @@ fn generate_list_and_response_models(
 ) -> (proc_macro2::TokenStream, proc_macro2::TokenStream) {
     // Generate List model
     let list_name = format_ident!("{}List", api_struct_name);
-    let raw_fields = extract_named_fields(input);
+    let raw_fields = field_analysis::extract_named_fields(input);
     let list_struct_fields = crate::codegen::models::list::generate_list_struct_fields(&raw_fields);
     let list_from_assignments =
         crate::codegen::models::list::generate_list_from_assignments(&raw_fields);
@@ -1088,12 +904,12 @@ pub fn entity_to_models(input: TokenStream) -> TokenStream {
         };
 
     // Extract fields and create field analysis
-    let fields = match extract_entity_fields(&input) {
+    let fields = match field_analysis::extract_entity_fields(&input) {
         Ok(f) => f,
         Err(e) => return e,
     };
-    let field_analysis = analyze_entity_fields(fields);
-    if let Err(e) = validate_field_analysis(&field_analysis) {
+    let field_analysis = field_analysis::analyze_entity_fields(fields);
+    if let Err(e) = field_analysis::validate_field_analysis(&field_analysis) {
         return e;
     }
